+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Bugfix - App User Management - Confirmation and User Role Changes
 * Shared Components
    * More versatile image component
+* Feature - User Management - fetch roles (api refactoring)
 
 
 ## 0.7.0